refactor(frontend): rename newRAS to RAS
also remove nanhu RAS, which is deprecated.
